kotlin/journal3: Explicitly model (im)mutability of moments
Create a dedicated interface to handle the mutability of moments. This
segregation would later be used to implement the automated sentiment
capture functionality.
