+++ app-kmm-journal3/src/commonMain/kotlin/com/hadisatrio/apps/kotlin/journal3/moment/fake/FakeMoment.kt
@@ -21,6 +21,7 @@ import com.benasher44.uuid.Uuid
 import com.chrynan.uri.core.Uri
 import com.hadisatrio.apps.kotlin.journal3.datetime.Timestamp
 import com.hadisatrio.apps.kotlin.journal3.datetime.UnixEpoch
+import com.hadisatrio.apps.kotlin.journal3.moment.EditableMoment
 import com.hadisatrio.apps.kotlin.journal3.moment.Moment
 import com.hadisatrio.apps.kotlin.journal3.sentiment.Sentiment
 import com.hadisatrio.apps.kotlin.journal3.token.TokenableString
@@ -30,7 +31,7 @@ import com.hadisatrio.libs.kotlin.geography.Place
 class FakeMoment(
     override val id: Uuid,
     private val group: MutableList<Moment>
-) : Moment {
+) : EditableMoment {
 
     private var isForgotten: Boolean = false
