+++ app-kmm-journal3/src/commonTest/kotlin/com/hadisatrio/apps/kotlin/journal3/moment/TimeRangedMomentsTest.kt
@@ -40,6 +40,7 @@ class TimeRangedMomentsTest {
     @BeforeTest
     fun `Init moments`() {
         origin.forEachIndexed { index, moment ->
+            moment as EditableMoment
             moment.update(LiteralTimestamp(currentInstant - index.days))
         }
     }
